Add support for updating environments lazily during incremental check
Summary:
Currently, to update key `k` on a dependency-tracked sharedmem table `T`, we need to go through the following steps:
1) Oldify the mapping for `k` in `T`. Let's assume that the oldified value is `v`.
2) Update `T` (i.e. insert new binding `k` -> `v'` into it)
3) Remove the old batch and compare `v` and `v'`. If they are the same, do not generate any triggers for `k` as nothing changes. Otherwise, add all dependencies of `k` into the trigger set.

Note that everything listed above happens in the master process, as it is the only place where we can do oldify/removal.

Things work well if the number of keys to update is small. But when the number gets large and when the size of the values becomes big, grabbing all corresponding `v`s and `v'`s from shared memory and comparing them become the bottleneck.

This diff adds the ability to "lazily" (or "pessimistically") update the environment incrementally, meaning that the incremental update would look like this:

1) Purge the mapping for `k` in `T`.
2) Pessimistically assume that the value of `k` would have changed, and add its dependencies to the trigger set. No value fetch nor a comparison of any sort is required.

This would still work thanks to lazy environment building: even though we do not compute the new binding for `k` during update, it will get computed in subsequent phase when the value of `k` gets queried.

The lazy update scheme significantly cut down the time of `Environment.update`, at the expense of potentially triggering more (spurious) keys to be re-analyzed in downstream tables. It's a tradeoff and the tradeoff doesn't always favor lazy update, so we'll need to keep the eager update and make it possible for the client to choose between the two.

I choose to place the knob as a required boolean value to the module signature of `Environment` -- when the flag is `false`, we do eager incremental update; otherwise we do lazy incremental update. In this diff, I put `false` everywhere to preserve the original semantics. We'll flip those switches as followups.

Reviewed By: mrkmndz

Differential Revision: D19349832

fbshipit-source-id: 2b146bf1ce3dbc1396e53eeb843c657d3e735103