Bypass land issues when landing global pyre configuration upgrade diff
Summary:
We definitely should not bypass land issues whenever we are touching non pyre-configuration files, but it looks like it's impossible to get auto land working for upgrade diffs if we block on this for the base global configuration upgrade. This diff will only ever touch files named `.pyre_configuration` or `.pyre_configuration.local`, so cannot break any code.

There are expected failures on this diff that we always need to bypass because projects that turn push-blocking on and sometimes still have type errors on master - when dealing with hundreds of configs this almost always is the case for at least one. Not sure if there is another way around this, besides resorting to manually shipping all upgrade diffs in the future.

Individual project upgrade diffs are also NFC, but they do go into actual python files and add comments. Those should/will continue to be blocked by any test or lint failures.

Reviewed By: grievejia

Differential Revision: D19518152

fbshipit-source-id: 464ab240999a5ac989283edb82ea11aeddd8db1f