introduce `analyze_assignment` abstraction to backwards analysis
Summary:
We were previously mechanically handling assignments (i.e. calling the underlying functions) for the `__getitem__` case in backward. Let's abstract things up a bit so that we don't have to re-write this logic in multiple places.

Concretely, fixes a bug where index taint would not get cleared on re-assignment for dictionaries, etc.

Reviewed By: dkgi

Differential Revision: D19539992

fbshipit-source-id: ec650a84e01a33c9e20359860144b246a6c507d3