pull out `FunctionDefinition` from UnannotatedGlobalEnvironment and purify collect_definitions
Summary: The goal here is to purify the `defines` query, which is currently not working when we have external files (they don't get added to the UnannotatedGlobalEnvironment for perf/memory reasons). This diff is part 1 of 2: Separating out the collect_defines functionality to have a pure part + the actual modification of the environment. A later diff will change `defines()` to use the pure function.

Reviewed By: grievejia

Differential Revision: D19548493

fbshipit-source-id: 560d8522e73857c22f95188e47a2b726446c248c