Adding mem dependency edges
Summary:
Currently whenever a value in shared memory changes, we trigger all of the dependencies.  These could include those that are only checking for existence of a value with Table.mem.
These can be avoided by separating those edges as being Mem edges and Get edges, and only triggering the mem edges when its existence changes, not its value
This isn't useful for the lazy environments because they are defined to always have values (which could be optionals).
However, the most relevant table which is Mem checked the most, UnannotatedGlobalEnvironment.ClassSummaries isn't lazy, and so we can use this Mem edge to save significant triggers.

Reviewed By: grievejia

Differential Revision: D19549330

fbshipit-source-id: 3c9f57a38e83658ee35f0c2e22f4a81e7ab38fe4