demangle nested callable names in issues
Summary:
We were not delocalizing nested functions, which meant that the external output was exposing Pyre's mangled names. Fix both the user-visible errors and the models/issues we output.

Before, in the test, the models would be reported for `$local_nested_callables?foo$inner` - we now report this as `nested_callables.foo.inner`. Since we qualify with the nested define, the names are unique unless you have multiple nested callables that share the same name in a given callable. In that case, we happen to be safe because Pyre's logic that fetches defines will only fetch the first define matching a name (for this exact reason)

Reviewed By: fahndrich

Differential Revision: D19592183

fbshipit-source-id: 5316642520445f310f658c32deb8904873be3608