Feed empty typeshed into client integration tests
Summary:
Let's not pass in any particular typeshed to these tests -- the content of typeshed shouldn't matter.

Also, if we're overriding the binary when running these tests from the Makefile, then they would fail if there's no version set underneath that override with which to look for the typeshed. We should avoid that.

Reviewed By: MaggieMoss

Differential Revision: D19558179

fbshipit-source-id: e94df202bba3297cb09373e6a294ad3509d713e2