Grab binary version from published hash
Summary:
Result of offline discussion --
There's a tradeoff on how much we should pull in from the configuration that sits on master / increasing dependence of the testing on the environment.
For now, just pull the binary version from a pyre invocation and hard-code any other configuration fields we care about.

This is better than using the pyre binary built on master due to easier repros (ready version hash), also to make testing a little more reliable & true to how the client would behave if we released it as-is at the moment of testing.

Reviewed By: grievejia

Differential Revision: D19453154

fbshipit-source-id: 4a284601282bc285f2c27113ca6136992aa43ab1