Support json output of servers list
Summary:
The output style is a common argument across pyre commands. Ran into the desire to get the output of `pyre servers` in a more structured way.

While I'm at it, this also modified pyre servers to use the same logging interface as other commands rather than `print`, and moved the testing away from mocking out the constructor of the Servers command. A lot of arguments that are valid on the Severs object are set during construction in its parent classes; migrated to passing in mock args / config like other commands do during testing.

Reviewed By: sinancepel

Differential Revision: D19626321

fbshipit-source-id: 98ef8cbd89d48644491fcc34d309df6576783a5f