Make Pysa integration tests print all differences instead of stopping on first one
Summary:
Our existing setup pre-emptively stops a test when it sees a difference in callgraphs and overrides, even if the models themselves might've changed. This means that once an existing test file is extended, you need to run the test 3 times to re-generate the new artifacts.

Change the setup so that all differences are written to the FS and reported to the user, so that we know that if we fix all the divergences in the test output, the test will succeed next time.

Reviewed By: fahndrich

Differential Revision: D19627568

fbshipit-source-id: 4f6c1157cc203561bf2f2a5882d9887cda59b0da