Sealing AnnotatedAttribute.t
Summary:
By sealing this, we ensure that we always go through the modified accessors on this type.
For example, annotation wraps the types with Awaitable if its an async attribute, which was only occurring in some of the usages of this type.
This also makes the injection of "late instantiation" more feasible.

Reviewed By: sinancepel

Differential Revision: D19635304

fbshipit-source-id: 239ca5f24a9aab1e33c10a56fb017cb4f6f93ebf