Clean up `Scheduler.t`
Summary:
A `Scheduler.t` can either be sequential, in which case everything is done in the master process and the number of workers is irrelevant; or it can be parallel, in which case we need to go through all the forking stuffs so workers can be set up. It can't be both, but the data structure we choose allows the "both" option.

Let's tighten up the types so we could be more mindful about where to fork and where not to fork.

Reviewed By: MaggieMoss

Differential Revision: D19647551

fbshipit-source-id: e40a5d377372709df53f32fd8e11c9103f6379b7