Hiding the attribute table inside of AttributeResolution
Summary:
As we evolve the storage representation of the attribute tables, it would be nice to clarify what the interface we intend to support is, abstracting the actual data representation.
It turns out we only ever use get_single_attribute, get_all_attributes, and get_all_attributes_names.

By hiding this, the next steps of late instantiation become easier.

Reviewed By: sinancepel

Differential Revision: D19648084

fbshipit-source-id: f71de96d6910974617577021e4dc04bb4d8365a9