Simplifying AnnotatedClass.overrides
Summary:
These two forms are actually equivalent due to the special casing on constraints when the instantiated type is Primitive parent.

Doing it all in one shot clarifies what we actually mean here, which is to find the attribute in a parent, with generics propagated.

Note that this creates a small regression in full check time, due to cache misses from that instantiated type.  That impact is removed by late instantiation.

Reviewed By: grievejia

Differential Revision: D19668655

fbshipit-source-id: 4b63c44c1dac5100845a8574027fa1821f4f8c94