Fix incorrect parenthesization in add_placeholder_stub_inheritances
Summary:
The purpose of add_placeholder_stub_inheritances is to augment a class that inherits from a class in a placeholder stub with all attributes, and a permissive constructor.  This is because for all we know the parent class has all the attributes in the world.  However, what we were doing was only adding those attributes if there was no constructor.  This bad behavior was implemented (it looks like by accident) by me in D18261571.

File this as reason #1003 why the if pred then unit_expression; syntax is dangerous :/

Reviewed By: grievejia

Differential Revision: D19673026

fbshipit-source-id: d0413d52cae2815c74821442c47504442abdad45