Nominally separate `InferenceError` from `AnalysisError`
Summary:
What this diff does: Add a module called `InferenceError`, and define it as `module InferenceError = AnalysisError`. The rest is to remove the alias of `Error = AnalysisError` in the analysis library, and deal with the consequences.

There are more info needed to be tracked for inference errors than those for analysis errors. Currently we confound the two and have to track everything in analysis errors. This is less than ideal as it bloats the  `AnalysisError` data structure, making us waste space as we store those errors into sharedmem.

This diff attempts at making a baby step towards separating the two. Granted, `InferenceError` is still pretty much an alias of `AnalysisError` in terms of implementation. But hey, at least those two are now nominally different, which I would call a good starting point!

Reviewed By: sinancepel

Differential Revision: D19633695

fbshipit-source-id: 3685c606912c55009c36d35c74c913618091e85b