Add standalone `lint` binary command
Summary:
I nuked the cli entry point to run additional checks in D18697230. Now that we are in a better position with respect to the type checker itself, it's time to bring it back. The reason why I choose to have a new command for it, as opposed to sharing the infra with `checkCommand`, is to avoid the `pyre infer` situation, where tightly coupled implementations made it really hard to evolve `pyre infer` and `pyre check` separately.

I'll tentatively call this command `pyre lint` but open to other suggestions on the name. Nothing has been implemented yet. This diff is just to make sure the command exists, so we don't need to do yet another binary upgrade later when we actually need it.

Reviewed By: dkgi

Differential Revision: D19670339

fbshipit-source-id: 2e8f40c716d39686b27800f46bb8be35d1bb66ab