Add standalone `infer` binary command
Summary:
This is absolute necessary to completely decouple `pyre infer` from `pyre check`.

Currently those two goes through the same backend command (i.e. `checkCommand`). This decision is precisely what makes infer and check so hard to separate in the past. To get around it we had to resolve to tons and tons of code duplications, which is bad.

Reviewed By: dkgi

Differential Revision: D19663968

fbshipit-source-id: 69e3510cd8e4998452c12fa6bbbc772052db63bb