support DictionaryKeys for abstract tree domains
Summary:
The idea of a dictionary key doesn't fit well with our current access path abstraction. The fundamental problem is that we want to be able to see tainted keys when `.keys()` is called on a dictionary or it is passed into a sink as is, but don't want to mark `d[unknown_key]` as being possible to taint as a key.

There were a few approaches to this problem.

1. Store the key and values of dictionaries as `"keys" -> KeyTaint, "values" -> regular taint. This certainly works, but is not ideal, as it'd double the depth of dictionaries and make the analysis more expensive.
2. Hack around, and use a convention of using a "$keys" field to store the key taint. This helps somewhat, but still runs into issues with regards to interactions with `[*]` - we'd need to add special logic to exclude `"$keys"` from star in joins as well, and this felt like it'd introduce tech debt and be hard to maintain (not to mention potential collisions with users choosing "$keys" as names).
3. Create a special label named `DictionaryKeys`, and write out its interactions with the other labels. This is the approach I went with.

All this diff does is add the label to the abstract tree domain, write out its interactions for join/leq/what have you, and add a few tests.

What I'd like advice on: Do the semantics make sense? I'm particularly not sure about what we should do when one of the <keys> fields is not present in LEQ.

Reviewed By: fahndrich

Differential Revision: D19644043

fbshipit-source-id: 0d22f9591cae2d1ba267283155300e9adef10086