Two-phase attribute table building
Summary:
This is Pt. I/II of "late instantiation"

One of the key problems of the attribute table is that it has a separate key/value for every combination of class name and instantiated type that it's ever been queried with.  This is necessary because the attributes contain the type of the attribute, which depends on the instantiated type.
What this diff accomplishes is disentangling the logic of attribute_table and create_attribute to the point where there are two distinct phases of creating an attribute table:
* collection and parsing, where we traverse the class hierarchy and parse the annotation expressions, interpret the metadata, etc.  This no longer requires having an instantiated type around
* instantiation: where we take the result of the first phase, and instantiate it into a full-fledged attribute with real types.

The impact of this is that we are wasting time doing things we largely have already done, and then filling up our memory with an ever increasing pile of very similar looking attribute tables.

The key insight of this diff is that these two units of work are fundamentally distinct, as long as you are crafty with what that intermediate representation looks like.  This means we can cache the results of the first phase, while deferring the second phase to when the attribute is actually requested.

This diff doesn't quite go that far, but instead just does the finicky work of splitting out these incredibly entangled phases.  This sets us up for the following diff, where we move phase two to be after the cache.

Things to look out for:
* I made AnnotatedAttribute.t a parametric type, to support the same structure containing both the instantiated and uninstantiated versions of the attribute.
* I made two different table structures for now.  This clearly could be a functor, but given that this will just exist for one diff (as we will no longer want the InstantiatedAttributeTable version after the next diff), I think its fine.
* I left some odd original names for things in instantiate_attribute in order to minimize the size of the changes to help the diff show up better.  I can clean it up in a bit.

Reviewed By: grievejia

Differential Revision: D19668009

fbshipit-source-id: 062c20ce68d4ce3a9e36d43ed499417749590eb0