Use `DictionaryKeys` in forward_analysis
Summary:
Pretty straightforward - prepend the "Keys" label to key taint in dictionaries & comprehensions.

Future work: Making the `.keys()` method return the key taint for dictionaries - right now, it's obscure, and will return the whole taint.

I don't think we have an equivalent in backwards analysis, as I can't imagine a situation in which a dictionary key is a sink.

Reviewed By: fahndrich

Differential Revision: D19651444

fbshipit-source-id: 3d2ed24beff424395084e63d54e7f449382ac7be