Cache last deleted file to handle back-to-back deletion and addition of a single file.
Summary:
Problem: When Vim edits a file, it moves the file to a swap file and moves it back. This looks like a "deletion" followed by an "addition". Pyre correctly updates on both. However, if there is a `pyre` request in between, it can return "could not find module foo" errors even though a `pyre` request one second later will show the right errors.

One reason for the race condition was that handling a new file requires a buck query, which can take a couple of seconds. This diff caches the last single-file deletion and bypasses the buck query if the next addition is for the same file.

Reviewed By: sinancepel

Differential Revision: D19646398

fbshipit-source-id: 378651f8f567090a88c569bf5771f1e621c40b6f