Refine final attributes of non local-finals
Summary:
This diff finally allows proper refinement on final attributes of non local-finals. Prior to this, Pyre had a much stricter rule for refinement; that is - we would only allow refinement of attributes that **strictly** came from a line of finals. However, this unfortunately created room for frustration among python devs due to the rigid behavior of our system.

This change is possible due to the introduction of `RefinementUnit`, which lays out the facilities to properly handle when we should or should not refine on the final attributes of local objects.

We also have to make sure to invalidate the refinement if the local is ever reassigned, which is trivially accomplished due to `RefinementUnit`'s nested structure of storing attributes.

The notable changes in this diff are 1) removing the constraint that an attribute that can be refined must come from a line of all finals and 2) properly invalidating the final attribute refinement should the local ever be reassigned (which is handled by `Resolution.unset_local`).

This is what we previously required for the program below to typecheck:
```
dataclass(frozen=True)
class Actor:
  name: Optional[str]
  def __init__(self, name: Optional[str]) -> None:
    self.name = name
def foo(a: Actor) -> Non
  a: Final[Actor] = a
  if a.name is not None:
    expects_str(a.name) # type-safe since a is a Final[Actor]
```
*note the strict requirement for `Final`*

Now:
```
dataclass(frozen=True)
class Actor:
  name: Optional[str]
  def __init__(self, name: Optional[str]) -> None:
    self.name = name
def foo(a: Actor) -> Non
  if a.name is not None:
    expects_str(a.name) # type-safe!
```
we no longer require the local to be a `Final`

... while properly invalidating the refinement if reassignment occurs:
```
dataclass(frozen=True)
class Actor:
  name: Optional[str]
  def __init__(self, name: Optional[str]) -> None:
    self.name = name
def foo(a: Actor) -> Non
  if a.name is not None:
    a = Actor(None)
    expects_str(a.name) # unsafe!
```

Reviewed By: mrkmndz

Differential Revision: D19665672

fbshipit-source-id: 11eb84965d05571651bc69d432f60ed45ed025f3