Allow the binary to accept extra feature flags
Summary: If `Yojson.Safe.Util.member` does not find a matching key it will return `Null`, this means that if a flag is removed on the client side it will default to being false on the binary. Looking at the pattern of the existing flags that makes sense to me, should a feature flag contain a string or another data property, we will need to update this `create` functionality.

Reviewed By: grievejia

Differential Revision: D19736492

fbshipit-source-id: 53cbf86f9d54c6b7102ea56449f6d7c37965b5aa