Purge the usage of `Annotation.is_global`
Summary:
Whether a variable is global or not should be a property of the variable, not a property of the annotation of the the variable. Therefore it does not make sense to keep a local/global flag on `Annotation.t`. Let's get rid of that.

We'll first get rid of all invocations to `Annotation.is_global` here. Followup diff will remove the flag completely.

Reviewed By: shannonzhu

Differential Revision: D19748424

fbshipit-source-id: a2408aba3815ab72a93741c0595c8cb1e93f4214