Changing implementation & messages of incompatible overload to no longer need define_location
Summary:
As described in the preceding diffs (D19753648 & D19747976), retaining locations in the values of shared memory tables is more problematic than it first appears.  A huge source of these locations right now is the define_locations we keep inside of callable types.  Because types are intrinsic to the value types of the attribute, parse annotation, and annotated global tables, we can't cut that value out, instead we need to actually remove the location.

The only reason that this field exists is to enable incompatible overload checking (introduced in D15824400).

To remove it, we need to come up with new algorithms for the existing checks that don't rely on locations. The key change I made was instead of only running the checks on the implementation function, and back-assigning the errors to the offending overload via define_location, I instead run the check once on each overload, for that particular overload and assign it to the location of that overload directly.  This also indirectly fixes a small issue with the existing implementation where errors on overloads were being silenced if the implementation wasn't annotated.  Furthermore, in order to eventually make the "offset pass" plan work, we will need to restrict the errors coming from checking a given function to be for lines inside that function, which the existing implementation did not follow.

The changes were:
* for IncompatibleOverload:ReturnType or Parameters, we previously folded through the overloads when checking the implementation.  Now for each overload we compare against the implementation.  The error message did not need to change, but it moved from being coalesced at the implementation to being located on each offending overload.
* for Unmatchable, instead of folding through the overloads like 1, 12, 123, 1234 on the implementation, we instead use the following trick:   we parse the overload we're working on currently then **find the overload within the list of overloads via Callable.equal_implementation**.  With it found, we can look at the preceding overloads in much the same fashion as before.  There are two behavioral changes that this entails:
  * The error messages change from "Incompatible overload [43]: The overloaded function `foo` on line 9 will never be matched.  The signature of overload on line 5 is the same or broader." to "Incompatible overload [43]: The overloaded function `foo` on line 9 will never be matched.  The signature `(bar: object, x: object) -> int` is the same or broader."  This is necessary because this new approach doesn't give us access to the locations of the other implementations, just the one we're currently working on
   * We no longer error on **exact** overload duplication, i.e. precisely same parameters and return type, since the search will detect the first one as the position for both.  I think this is a pretty uncommon, and fairly benign issue, but if we want to cover it, we could error on the implementation when it has exact duplicates in its overload list.  Not sure if I think that's worth it.

Reviewed By: MaggieMoss

Differential Revision: D19761501

fbshipit-source-id: 082acbca1f88825032d53aa6e6618664c64d169e