removing lookup fallback to define_location
Summary: The only other consumer of define_location was go to definition.  Given that this is currently disabled, this is a very low impact loss (especially compared to why we're removing this).  Furthermore, this only works for callables, and a robust go to definition system would use a robust lookup into the unannotatedglobalenvironment to handle all kinds of attributes.

Reviewed By: sinancepel, grievejia

Differential Revision: D19763305

fbshipit-source-id: 5be2a1652660210a22519d66898f0bc2c79722c9