Splitting off global locations
Summary: Now that we've cleaned up our act with the attribute table, the other big offender with location-based triggers is the annotated global table.  This is another thing whose semantics are entirely location free, except for one or two sites where we want the location.  I resolve this by building (yet another) environment between the annotated global environment and the type environment, just to extract (and cache) those locations.  I'm not sure caching these is really necessary, but it was easy and I couldn't come up with a reason not to.  This also potentially prefigures an opportunity for the unannotated global environment.

Reviewed By: grievejia

Differential Revision: D19776968

fbshipit-source-id: 7a9ca20d556ab100692a1d6cca4a0a1318e676e0