Add framework to collect non-final attributes from source files
Summary:
This change implements a basic framework for amassing all guaranteed non-final attributes on a class level. This is done by checking all assignments after initialization (i.e. visitor must not be in a constructor). This will help us identify which fields can be marked final, to drive adoption of final attribute refinement.

Note: this currently only collects the attributes belonging to the same class. Work has to be done to collect fields belonging to outside classes - which will inevitably necessitate the resolution.

Reviewed By: grievejia

Differential Revision: D19755973

fbshipit-source-id: 50f9af003d446ea1c6efa967d94d4a9c5715308a