Do not call into `TypeCheck.State.errors` from server queries
Summary:
# Context

My high-level goal here is to take error computation out of resolution fixpoint computation. The idea is that the `State` module should only be concerned with types and not with anything that's related to errors.

Unfortunately, there are existing usages that depends on `State.errors` directly. I'll need to get rid of them first if I ever want to do anything about states and errors.

# This Diff

One of the external usage of `State.errors` comes from the server query where the type of a given expression is requested. It uses `State.errors` because it wants to know whether the given expression itself is well-typed.

Ideally, we'd want a dedicated API from either `TypeCheck` or `Resolution` for this use case (e.g. return an optional from `Resolution.resolve`) and redirect the query there. But since I figured that the type query is not a critical part of the system and the semantics of the query is not strictly defined as well, it shouldn't be too big of a deal if I just break the functionality of error-awareness and redefine the query to mean "return the type of the expression regardless of type errors".

After all, building a dummy state and calling into `State.forward_expression`/`State.errors` directly is just code duplication given that we already have `Resolution.resolve` doing exactly the same thing.

Reviewed By: sinancepel

Differential Revision: D19792189

fbshipit-source-id: 653d880941a827c928952b36a07fc5f1bf4b0b3f