Hide `State.errors` from the `TypeCheck` module
Summary:
# Context

My high-level goal here is to take error computation out of resolution fixpoint computation. The idea is that the `State` module should only be concerned with types and not with anything that's related to errors.

# This diff

With D19792189 and D19798709, we've removed all external usages of `TypeCheck.State.errors` from production. Let's hide that API from `TypeCheck` so future changes on that function will only affect one single module, allowing me to move faster.

Remaining invocations of `State.errors` are all in tests. As mentioned before, the `State` module should not be concerned about errors. So I've removed the assertion on error messages for the majority of state-related tests. I doubt we'll lose much in test coverage though, as error changes are already guarded by a large number of integration tests.

Reviewed By: dkgi

Differential Revision: D19799677

fbshipit-source-id: 67fefb0fe24326487efe2dcab907eee964c48a00