Parallelizing the detection of changes to dependency tracked memory
Summary:
A core element of our approach to dependency tracking is "stops", i.e. an element may be triggered to be recalculated, but if the value hasn't been changed, then we need not trigger the things that depend on it.

For large updates, a significant amount of time was spent (serially) getting the old and new values from shared memory, and then serially comparing them.  We can save some time in some heavily loaded scenarios by doing so in parallel.

Reviewed By: grievejia

Differential Revision: D19508840

fbshipit-source-id: 2d6ad54ea1782e361751211638be66113927d819