Implementing non-user accessible parameter specification concatenation
Summary:
I realized that a consequence of two properties of PEP 612 is that prepending anonymous arguments to a paramspec is permitted.

Those properties are:

>
One additional form that we want to support is functions that pass only a subset of their arguments on to another function. To avoid shadowing a named or keyword only argument in the ParameterSpecification we require that the additional arguments be anonymous arguments that precede the *args and *kwargs
```
def call_n_times(
    __f: Callable[TParams, None],
    __n: int,
    *args: TParams.args,
    **kwargs: TParams.kwargs,
) -> None:
    for x in range(__n);
        __f(*args, **kwargs)
Backwards Compatibility
```

and

> Just as with normal TypeVars, ParameterSpecifications can be used to make generic classes as well as generic functions. These ought to be able to be mixed with normal TypeVars. This should also be made to work with protocols in the same manner.

This together allows you to make
```

      class CallableReturningInt(Protocol[TParams]):
        def __call__(__self, __f: int, *args: TParams.args, **kwargs: TParams.kwargs) -> int:
          ...
```

Which is a protocol that effectively prepends an anonymous int argument to a paramspec.

This diff supports this behavior, and paves the way for a user-accessible concatenation

Reviewed By: dkgi

Differential Revision: D19538934

fbshipit-source-id: 5982c4448c01a4c03114fd49d445181ec1deb572