refactor lookups to re-compute comprehension bindings on demand
Summary:
Context: I discovered a flaw today in our fixpoint logic - we dynamically compute the resolution
fixpoint during every iteration of type checking, and keep both the expression + statement
resolution fixpoint around. However, our fixpoint doesn't take the resolution fixpoint into account
(and likely can't without a significant performance hit), and resolution fixpoints corresponding to
non-final states were exported. This led to correctness bugs in Pysa,where we rely on the resolution
fixpoint for type info.

As a result, I plan to work on moving the computation of the resolution fixpoint completely out of
the regular fixpoint, and instead compute it once a fixpoint is reached per basic block.

Where this diff comes in: To accomplish the above, we need to not compute the resolution fixpoint
for each expression. The only consumer of the resolution fixpoint's expressions is lookups, so
instead re-compute the comprehension resolution as needed. As a bonus, we now properly refine
conditions ;)

Reviewed By: grievejia

Differential Revision: D19849664

fbshipit-source-id: b660ac3699b5c267f51ff1b0cf5f8cb184d1714c