remove expressions from local annotation map
Summary:
The motivation for having expressions in local annotation maps was presented in D18979747.
A few of the statements there are no longer the case (we *can* nowadays reasonably invoke
forward_statement from the lookup logic), so we can dynamically recompute these comphrehension types,
and indeed do in the diff below. This is thus now dead code :)

Reviewed By: grievejia

Differential Revision: D19849952

fbshipit-source-id: 662971decb77dbc41967d6ef0c78dcb282110e7a