A more scalable implementation of initialization checking
Summary:
The diffs up to this point were more or less the low-hanging fruit of eliminating the usages of class summaries.  From here on in they will need to be a bit more involved.

This diff tackles one of the most important users of full class summaries in typeCheck, initialization checking.
Big picture, the idea behind initialization checking is that we want to error on concrete classes which do not initialize all of the properties they claim to have.

The current implementation works something like this:
for every class we retrieve its class summary and its attribute table, and then create the implicit attribute table from the class summary.
For each attribute, we check whether its marked as defined in the table or present in the implicit attribute table.
If its neither of these (and also isn't "dynamically initialized") then we show an error.

The reason that the class summary is needed here is to make the implicit attribute table, which is there to support the following scenario:
```
class Foo:
   x: int # the "explicit" attribute is preferred to the implicit (from the constructor), marked as uninitialized
   def __init__(self, x: int) ->  None:
        self.x = x # this "implicit" attribute is initialized, but we never look at that.
```

To patch that behavior, we currently just look for the presence of a name in the implicit attribute map. This means we currently do not error in this scenario.

```
class Foo:
   def __init__(self) ->  None:
        self.x : int # this "implicit" attribute is uninitialized, but we never look at that.
```

This diff fixes that behavioral oddity while also making this more sustainable by no longer requiring the implicit attribute table, and thus not needing the class summary.

The core problem here is that in our merges of the various attribute tables, we select the attribute in a priority order, with explicit coming before implicit.  This is good because it preserves manual annotations and other properties, but is bad in this case because we lose the fact from the implicit attribute that it is in fact instantiated.  This leads to the annotated attributes being spuriously marked as uninstantiated.

To fix this I introduce a new states for the initialization of an annotated attribute, Explicitly instantiated, implicitly instantiated and not instantiated.  This is defined in terms of a new semantic of the merges, instead of preserving the prioritized attribute wholesale, we collect all of the values in a list.  I also had to separately mark nested classes since they don't have an expression value, but are in fact instantiated.

One note:
The reason I had to distinguish between explicitly and implicitly instantiated attributes, which requires that list, rather than just binary instantiatiated-ness, which would only require prioritizing values over non-values, is that we want to forbid this case:
```
class Foo:
   x: Final[int] = 1
   def __init__(self, x: int) -> None:
       self.x = x
```

This needs to be locally (in the init) distinguishable from this
```
class Foo:
   x: Final[int]
   def __init__(self, x: int) -> None:
       self.x = x
```
which we want to support.  This requires a distinction in the kinds of instantiation.

Reviewed By: dkgi

Differential Revision: D19786168

fbshipit-source-id: 9519e231d44dfa7eb4818d8c701b6f112aa4fdbc