Take `check_return` out of typechecking state
Summary:
The `check_return` field in state is computed in `State.initial`, but not used until later while we are validating return types. This kind of dependency is bad for my purpose -- I'd prefer the `check_return` field to be computable *when* we are validating return types, not *before* it. Ideally, the only dependency between typing and the error emission should be the resolution fixpoint. Throwing additional flags like `check_return` into the picture is both inconvenient and unnecessary.

This diff makes it such that the `check_return` flag is computed inside the `validate_return` function. Since we don't have the freshly parsed annotation available in `validate_return`, I have to make a small semantics change there. Previously, we check the return type if the parsed return annotation contains `Top`. This doesn't seem to be a sensible choice to me: if we return `int` in a function whose return annotation is `List[unknown_type]`, we'll suppress the error there. I changed it such that we only perform error suppression when we don't have a return annotation specified.

Reviewed By: dkgi

Differential Revision: D19827388

fbshipit-source-id: cfc6ff56ef55b572d6c6b81bd54eb37f2b2d1330